Smarter Instance Support
Add InstanceAdapter class which deals with keeping track of a
particular document's values.  Change type handling back to original
implementation having a type object handle a single instance of that
type on a particular document instnace.

Start clean-up of DesignDocument in an effort to move all of that
functionality to the class-level.