EmbededModel
Break out embeded documents by allowing a class to be passed to the
embeded method on the dsl which should be a Loveseat::EmbededModel.
EmbededModel exists as a wrapper to the uglier version using Model and
setuping the model as abstract. This also removes the persistence
methods that exist on Model from an embeded object.